[#update] 몬스터 생성구조 인터페이스화 및 웨이브 시스템 변경

+++ Assets/Scripts/DamageTextManager.cs
@@ -26,7 +26,7 @@ public class DamageTextManager : MonoBehaviour
         }
     }
 
-    public void PrintText(Vector3 pos, int damage, int dir)
+    public void PrintText(Vector3 pos, float damage, int dir)
     {
         var text = textPool.Pop();
         text.gameObject.SetActive(true);
